fix dtype of Normal.evalf for 1d argument
The `evaluable.Normal` class has a fast implementation of `evalf` for
one-dimensional geometries. This implementation, however, returns an array with
the same dtype as the argument of `evalf`, the local gradient of the geometry,
which could be different from the dtype announced by `evaluable.Normal`:
`float`. This patch fixes the problem by adding an explicit cast to a `float`.

+++ nutils/evaluable.py
@@ -1088,7 +1088,7 @@ class Normal(Array):
   def evalf(self, lgrad):
     n = lgrad[...,-1]
     if equalindex(n.shape[-1], 1): # geom is 1D
-      return numpy.sign(n)
+      return numpy.sign(n).astype(float)
     # orthonormalize n to G
     G = lgrad[...,:-1]
     GG = numeric.contract(G[...,:,_,:], G[...,:,:,_], axis=-3)
